requirement.txt
antlr4-python3-runtime==4.9.3
asgiref==3.7.2
blinker==1.8.2
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
distlib==0.3.8
Django==4.2.5
docutils==0.21.2
filelock==3.16.1
firebase-admin==6.6.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gcloud==0.18.3
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
hydra-core==1.3.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.5
msgpack==1.1.0
mysql-connector-python==9.0.0
numpy==1.26.4
oauth2client==4.1.3
omegaconf==2.3.0
packaging==24.0
pillow==10.4.0
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.29.0rc1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
pypiwin32==223
Pyrebase4==4.8.0
python-jwt==4.1.0
pywin32==307
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.20
virtualenv==20.26.6
Werkzeug==3.0.4
